Collect statistics for upstream_fair module
The shared memory is accessed without any locks. The number of
requests is updated atomically and the last activity time uses
a rotating set of slots a'la ngx_times.c

Note: the numer of time slots may need to be increased under heavy
load, but on the other hand, corrupted reads should lead in the
worst case to non-optimal load balancing. This needs further testing.